+++ backend/internal/storage/friend/storage.go
@@ -16,7 +16,7 @@ func New(dbRouter infra_database.DBRouter) *FriendStorage {
 	}
 }
 
-func (s *FriendStorage) AddFriend(ctx context.Context, userID, friendID int) error {
+func (s *FriendStorage) AddFriend(ctx context.Context, userID, friendID int64) error {
 	_, err := s.dbRouter.Exec(ctx,
 		`INSERT INTO follows (user_id, friend_id) 
 			VALUES ($1,$2) ON CONFLICT DO NOTHING`,
@@ -26,7 +26,7 @@ func (s *FriendStorage) AddFriend(ctx context.Context, userID, friendID int) err
 	return err
 }
 
-func (s *FriendStorage) DeleteFriend(ctx context.Context, userID, friendID int) error {
+func (s *FriendStorage) DeleteFriend(ctx context.Context, userID, friendID int64) error {
 	_, err := s.dbRouter.Exec(ctx,
 		`DELETE FROM follows WHERE user_id = $1 AND friend_id = $2`,
 		userID, friendID,
